+++ R/boot.LR.R
@@ -28,7 +28,7 @@
 #' # standard LR test
 #' anova(mod1, mod2)
 #'
-#' # boostrap LR test (run in parallel to save time)
+#' # bootstrap LR test (run in parallel to save time)
 #' mirtCluster()
 #' boot.LR(mod1, mod2, R=200)
 #'